+++ src/clojure/nrepl/middleware/pr_values.clj
@@ -7,16 +7,16 @@
   (:import
    nrepl.transport.Transport))
 
-(defn- default-renderer
-  "Uses `print-dup` or `print-method` to render a value to a string."
+(defn- default-printer
+  "Uses `print-dup` or `print-method` to print a value to a string."
   [v opts]
   (let [printer (if *print-dup* print-dup print-method)
         writer (java.io.StringWriter.)]
     (printer v writer)
     (str writer)))
 
-(defn- resolve-renderer
-  "Resolve a namespaced symbol to a rendering var. Returns the var or nil if
+(defn- resolve-printer
+  "Resolve a namespaced symbol to a printer var. Returns the var or nil if
   the argument is nil or not resolvable."
   [var-sym]
   (when-let [var-sym (and var-sym (symbol var-sym))]
@@ -24,13 +24,13 @@
       (require (symbol (namespace var-sym)))
       (resolve var-sym)
       (catch Exception ex
-        (log/warn "Couldn't resolve rendering function" var-sym)
+        (log/warn "Couldn't resolve printer function" var-sym)
         nil))))
 
-(defn- rendering-transport
-  "Wraps a `Transport` with code which renders the value of messages sent to
+(defn- printing-transport
+  "Wraps a `Transport` with code which prints the value of messages sent to
   it using the provided function."
-  [^Transport transport render-fn opts]
+  [^Transport transport print-fn opts]
   (reify Transport
     (recv [this]
       (.recv transport))
@@ -41,19 +41,19 @@
              (if (and (string? (:value resp)) (:printed-value resp))
                (dissoc resp :printed-value)
                (if-let [[_ v] (find resp :value)]
-                 (assoc resp :value (str/trim-newline (render-fn v opts)))
+                 (assoc resp :value (str/trim-newline (print-fn v opts)))
                  resp)))
       this)))
 
 (defn pr-values
   "Middleware that returns a handler which transforms any `:value` slots in
   messages sent via the request's `Transport` to strings via the provided
-  `:renderer` function, delegating all actual message handling to the provided
+  `:printer` function, delegating all actual message handling to the provided
   handler.
 
-  If no custom renderer is set, this falls back to using `print-dup` or
+  If no custom printer is set, this falls back to using `print-dup` or
   `print-method`. The function will be called with the value and any resolved
-  `:render-options` from the message.
+  `:print-options` from the message.
 
   Requires that results of eval operations are sent in messages in a
   `:value` slot.
@@ -63,9 +63,9 @@
   evaluation contexts to produce printed results in `:value` if they so choose,
   and opt out of the printing here."
   [handler]
-  (fn [{:keys [^Transport transport renderer render-options] :as msg}]
-    (let [render-fn (or (resolve-renderer renderer) default-renderer)
-          transport (rendering-transport transport render-fn render-options)]
+  (fn [{:keys [^Transport transport printer print-options] :as msg}]
+    (let [print-fn (or (resolve-printer printer) default-printer)
+          transport (printing-transport transport print-fn print-options)]
       (handler (assoc msg :transport transport)))))
 
 (set-descriptor! #'pr-values